Move cops related to `sig` enforcement to `signatures/`
