+++ lib/rubocop/cop/sorbet/signatures/allow_incompatible_override.rb
