+++ lib/rubocop/cop/sorbet/signatures/checked_true_in_signature.rb
