+++ spec/cop/sorbet/signatures/allow_incompatible_override_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-require_relative '../../../lib/rubocop/cop/sorbet/allow_incompatible_override'
+require_relative '../../../../lib/rubocop/cop/sorbet/signatures/allow_incompatible_override'
 
 RSpec.describe(RuboCop::Cop::Sorbet::AllowIncompatibleOverride, :config) do
   subject(:cop) { described_class.new(config) }
