+++ spec/cop/sorbet/signatures/checked_true_in_signature_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-require_relative '../../../lib/rubocop/cop/sorbet/checked_true_in_signature'
+require_relative '../../../../lib/rubocop/cop/sorbet/signatures/checked_true_in_signature'
 
 RSpec.describe(RuboCop::Cop::Sorbet::CheckedTrueInSignature, :config) do
   subject(:cop) { described_class.new(config) }
